fix: move convert_ts_to_int_timestamp
